+++ src/test/java/org/schakalacka/java/raytracing/geometry/objects/SphereTest.java
@@ -116,15 +116,15 @@ public class SphereTest {
     @Test
     void normalNotOnAxis() {
         var sphere = new Sphere();
-        var normal = sphere.localNormalVectorAt(Tuple.point((float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3)));
+        var normal = sphere.localNormalVectorAt(Tuple.point( (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3)));
 
-        assertEquals(Tuple.vector((float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3)), normal);
+        assertEquals(Tuple.vector( (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3)), normal);
     }
 
     @Test
     void normalIsAlreadyNormalized() {
         var sphere = new Sphere();
-        var normal = sphere.localNormalVectorAt(Tuple.point((float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3), (float) (Math.sqrt(3) / 3)));
+        var normal = sphere.localNormalVectorAt(Tuple.point( (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3),  (Math.sqrt(3) / 3)));
 
         var normalized = normal.normalize();
         assertNotSame(normal, normalized);
